refactor(stats): déplacer la vue StatistiquesPageView dans l'application forum_stats
