+++ lacommunaute/pages/urls.py
@@ -15,7 +15,6 @@ app_name = "pages"
 
 urlpatterns = [
     path("", views.HomeView.as_view(), name="home"),
-    path("statistiques/", views.StatistiquesPageView.as_view(), name="statistiques"),
     path("accessibilite/", views.accessibilite, name="accessibilite"),
     path("mentions-legales/", views.mentions_legales, name="mentions_legales"),
     path("politique-de-confidentialite/", views.politique_de_confidentialite, name="politique_de_confidentialite"),
