+++ lacommunaute/templates/partials/footer.html
@@ -129,7 +129,7 @@
                             <a href="{% url 'pages:politique_de_confidentialite' %}" aria-label="Politique de confidentialité">Politique de confidentialité</a>
                         </li>
                         <li>
-                            <a href="{% url 'pages:statistiques' %}">Statistiques</a>
+                            <a href="{% url 'forum_stats:statistiques' %}">Statistiques</a>
                         </li>
                     </ul>
                     <p>
